Agregado el diagrama de caja y bigotes. Se actualiza la base de datos, pero será editada directamente en la web. Se hará una edición especial al diseño para ver si funciona. Si funciona hago otro commit

+++ Views/ingresarDatos.php
@@ -54,7 +54,7 @@
                 var c = parseInt($("#txtnclass").val());
                 var li = parseInt($("#txtli").val());
                 var w = parseInt($("#txtamp").val());
-                var d = parseInt($("#txtList").val());
+                var d = parseFloat($("#txtList").val());
                 var r = w * c;
                 if((d-li)/w <= c && d-li >= 0) {
                     $("#txtLD").val($("#txtLD").val() + (($("#txtLD").val().length > 0)? "\n" : "") + d);
